Break parsing code + types into types.{c,h}
