+++ include/criteria.h
@@ -52,9 +52,6 @@ void destroy_criteria(struct mako_criteria *criteria);
 bool match_criteria(struct mako_criteria *criteria,
 		struct mako_notification *notif);
 
-bool parse_boolean(const char *string, bool *out);
-bool parse_urgency(const char *string, enum mako_notification_urgency *out);
-
 bool parse_criteria(const char *string, struct mako_criteria *criteria);
 bool apply_criteria_field(struct mako_criteria *criteria, char *token);
